backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
